move the smoke-test harnesses out of main.cpp into SmokeHarness.cpp
A verbatim move: the 39 --*-smoke-test / --*-repro branches and their
helpers become smoke::dispatch(Context&), which main() calls before its
own option chain; the loopback server main() stopped after exec() lives
in the Context and smoke::shutdown() stops it. Both compile only with
AMREXPLORER_BUILD_TESTS, as before. main.cpp goes from 4,062 lines to
476, all production.
